requirements.txt
composabl
flask[async]
numpy
Jinja2
kafka-python
